# web/public/docs/api/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "DIY Accounting Submit API",
    "description" : "DIY Accounting Submit API documentation",
    "version" : "1.0.0",
    "termsOfService" : "https://submit.diyaccounting.co.uk/terms.html",
    "contact" : {
      "email" : "admin@diyaccounting.co.uk"
    },
    "license" : {
      "name" : "PolyForm Internal Use License 1.0.0",
      "url" : "https://github.com/diy-accounting-uk/submit.diyaccounting.co.uk/blob/main/LICENSE"
    },
    "x-authentication-guide" : {
      "description" : "This API is protected by Amazon Cognito. To use the API:\n\n1. **Via Website**: If you are logged into the DIY Accounting Submit website, you can access the Swagger UI directly and API calls will use your existing session cookies.\n\n2. **External Access**: \n   - First complete the Cognito hosted UI OAuth flow to get an authorization code\n   - Exchange that code for an access token by calling `/api/v1/cognito/token`\n   - Include the token in the `Authorization: Bearer <token>` header\n\n3. **For HMRC Integration**: Some endpoints require HMRC authorization tokens obtained via `/api/v1/hmrc/token`\n\n**Note**: All endpoints require proper CORS headers and are accessible from the same domain as the main application."
    }
  },
  "servers" : [ {
    "url" : "${env.DIY_SUBMIT_APEX_URL}api/v1/",
    "description" : "DIY Accounting Submit API documentation"
  } ],
  "paths" : {
...
    "/hmrc/token" : {
      "post" : {
        "summary" : "Exchange HMRC authorization code for access token",
        "description" : "Exchanges an HMRC authorization code for an access token",
        "operationId" : "exchangeHmrcToken",
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "code" : {
                    "type" : "string",
                    "example" : "abc123"
                  }
                }
              },
              "example" : {
                "code" : "abc123"
              }
            }
          }
        },
        "tags" : [ "HMRC" ],
        "security" : [ {
          "CognitoAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "HMRC token exchanged successfully"
          }
        }
      }
    },
    "/pass" : {
      "get" : {
        "summary" : "Validate a pass code",
        "description" : "Checks if a pass code is valid and returns its details",
        "operationId" : "validatePass",
        "parameters" : [ {
          "name" : "code",
          "in" : "query",
          "required" : true,
          "description" : "The four-word pass code to validate",
          "schema" : {
            "type" : "string"
          }
        } ]
      },
      "post" : {
        "summary" : "Redeem a pass code",
        "description" : "Redeems a pass code and grants the associated bundle to the authenticated user",
        "operationId" : "redeemPass"
      }
    },
...
    "/bundle/{id}" : {
      "delete" : {
        "summary" : "Delete bundle by id",
        "description" : "Deletes a bundle for the authenticated user using a path parameter",
        "operationId" : "deleteBundleById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "description" : "The bundle id (or name) to delete",
          "schema" : {
            "type" : "string"
          }
        } ],
        "tags" : [ "Account" ],
        "security" : [ {
          "CognitoAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "Bundle deleted successfully"
          },
          "202" : {
            "description" : "Request accepted for processing (async)"
          }
        }
      }
    },
...
    "/hmrc/vat/return" : {
      "post" : {
        "summary" : "Submit VAT return to HMRC",
        "description" : "Submits a VAT return to HMRC on behalf of the authenticated user",
        "operationId" : "submitVatReturn",
        "parameters" : [ {
          "name" : "Gov-Test-Scenario",
          "in" : "header",
          "required" : false,
          "description" : "HMRC sandbox test scenario",
          "schema" : {
            "type" : "string",
            "example" : "QUARTERLY_ONE_MET"
          }
        }, {
          "name" : "runFraudPreventionHeaderValidation",
          "in" : "query",
          "required" : false,
          "description" : "When true, validates HMRC Fraud Prevention Headers",
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "vatNumber" : {
                    "type" : "string",
                    "example" : "983238295"
                  },
                  "periodKey" : {
                    "type" : "string",
                    "example" : "24A1"
                  },
                  "vatDue" : {
                    "type" : "number",
                    "format" : "float",
                    "example" : 2400.0
                  },
                  "accessToken" : {
                    "type" : "string",
                    "example" : "HMRC_ACCESS_TOKEN"
                  }
                }
              },
              "example" : {
                "vatNumber" : "983238295",
                "periodKey" : "24A1",
                "vatDue" : 2400.0,
                "accessToken" : "HMRC_ACCESS_TOKEN"
              }
            }
          }
        },
        "tags" : [ "HMRC" ],
        "security" : [ {
          "HmrcAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "VAT return submitted successfully"
          }
        }
      },
      "get" : {
        "summary" : "Get submitted VAT returns from HMRC",
        "description" : "Retrieves previously submitted VAT returns from HMRC for the authenticated user",
        "operationId" : "getVatReturns",
        "parameters" : [ {
          "name" : "periodKey",
          "in" : "path",
          "required" : true,
          "description" : "The VAT period key to retrieve",
          "schema" : {
            "type" : "string",
            "example" : "24A1"
          }
        }, {
          "name" : "vrn",
          "in" : "query",
          "required" : true,
          "description" : "VAT registration number (9 digits)",
          "schema" : {
            "type" : "string",
            "example" : "983238295",
            "pattern" : "^\\d{9}$"
          }
        }, {
          "name" : "Gov-Test-Scenario",
          "in" : "query",
          "required" : false,
          "description" : "HMRC sandbox test scenario",
          "schema" : {
            "type" : "string",
            "example" : "QUARTERLY_ONE_MET"
          }
        }, {
          "name" : "runFraudPreventionHeaderValidation",
          "in" : "query",
          "required" : false,
          "description" : "When true, validates HMRC Fraud Prevention Headers",
          "schema" : {
            "type" : "string"
          }
        } ],
        "tags" : [ "HMRC" ],
        "responses" : {
          "200" : {
            "description" : "VAT return submitted successfully"
          }
        }
      }
    },
...
    "/hmrc/vat/obligation" : {
      "get" : {
        "summary" : "Get VAT obligations from HMRC",
        "description" : "Retrieves VAT obligations from HMRC for the authenticated user",
        "operationId" : "getVatObligations",
        "parameters" : [ {
          "name" : "vrn",
          "in" : "query",
          "required" : true,
          "description" : "VAT registration number (9 digits)",
          "schema" : {
            "type" : "string",
            "example" : "983238295",
            "pattern" : "^\\d{9}$"
          }
        }, {
          "name" : "from",
          "in" : "query",
          "required" : false,
          "description" : "From date in YYYY-MM-DD format",
          "schema" : {
            "type" : "string",
            "example" : "2025-01-01"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : false,
          "description" : "To date in YYYY-MM-DD format",
          "schema" : {
            "type" : "string",
            "example" : "2025-03-31"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "description" : "Obligation status: O (Open) or F (Fulfilled)",
          "schema" : {
            "type" : "string",
            "enum" : "O,F",
            "example" : "O"
          }
        }, {
          "name" : "Gov-Test-Scenario",
          "in" : "query",
          "required" : false,
          "description" : "HMRC sandbox test scenario",
          "schema" : {
            "type" : "string",
            "example" : "QUARTERLY_ONE_MET"
          }
        }, {
          "name" : "runFraudPreventionHeaderValidation",
          "in" : "query",
          "required" : false,
          "description" : "When true, validates HMRC Fraud Prevention Headers",
          "schema" : {
            "type" : "string"
          }
        } ],
        "tags" : [ "HMRC" ],
        "security" : [ {
          "CognitoAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "Request completed successfully"
          }
        }
      }
    },
...
    "/hmrc/receipt" : {
      "get" : {
        "summary" : "Retrieve stored receipts",
        "description" : "Retrieves previously stored receipts for the authenticated user",
        "operationId" : "getReceipts",
        "parameters" : [ {
          "name" : "name",
          "in" : "query",
          "required" : false,
          "description" : "Receipt file name including .json",
          "schema" : {
            "type" : "string",
            "example" : "2025-03-31-123456789012.json"
          }
        }, {
          "name" : "key",
          "in" : "query",
          "required" : false,
          "description" : "Full DynamoDB Item key",
          "schema" : {
            "type" : "string",
            "example" : "receipts/00000000-0000-0000-0000-000000000000/2025-03-31-123456789012.json"
          }
        } ],
        "tags" : [ "HMRC" ],
        "security" : [ {
          "CognitoAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "Receipts retrieved successfully"
          }
        }
      }
    },
...
    "/hmrc/receipt/{name}" : {
      "get" : {
        "summary" : "Retrieve a stored receipt by name",
        "description" : "Retrieves a specific stored receipt for the authenticated user by file name",
        "operationId" : "getReceiptByName",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "description" : "The receipt file name including .json",
          "schema" : {
            "type" : "string",
            "example" : "2025-03-31-123456789012.json"
          }
        } ],
        "tags" : [ "HMRC" ],
        "security" : [ {
          "CognitoAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "Request completed successfully"
          }
        }
      }
    },
...
  },
  "components" : {
    "securitySchemes" : {
      "CognitoAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT",
        "description" : "Cognito JWT token"
      },
      "HmrcAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "description" : "HMRC OAuth2 token"
      }
    }
  }
}